requirements.txt
Django==1.3.1
PIL==1.1.7
distribute==0.6.21
django-filebrowser==3.4.0
django-grappelli==2.3.7
django-guardian==1.0.3
django-tagging==0.3.1
django-tagging-autocomplete==0.3.1
django-tinymce==1.5.1b2
feedparser==5.1.1
reportlab==2.5
tagging==0.2.1
wsgiref==0.1.2
